HS-231/back/feature/word-game/checkAnswer 사용자는 낱말 게임 문제의 답안을 input text에 타이핑해서 제출할 수 있다.

+++ backend/src/main/java/com/hans/hans/domain/wordgame/dto/WordGameAnswerResponseDto.java
@@ -10,11 +10,11 @@ public class WordGameAnswerResponseDto {
 
     private Long roomSequence;
     private Map<String, Long> correctPlayers;
-    private int point;
+    private Long point;
     private String answer;
 
     @Builder
-    public WordGameAnswerResponseDto(Long roomSequence, Map<String, Long> correctPlayers, int point, String answer){
+    public WordGameAnswerResponseDto(Long roomSequence, Map<String, Long> correctPlayers, Long point, String answer){
         this.roomSequence = roomSequence;
         this.correctPlayers = correctPlayers;
         this.point = point;
